requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
flask-marshmallow==1.2.1
marshmallow==3.23.2
marshmallow-sqlalchemy==1.4.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Werkzeug==3.1.3
psycopg2-binary==2.9.11
python-dotenv==1.0.1
gunicorn==23.0.0
